sources(curl): error if curl exists 0 but there are downloads left
As part of the investigation of the CI failure in
https://github.com/osbuild/osbuild-composer/pull/4247
we noticed that curl can return a return_code of `0` even
when it did not downloaded all the urls in a `--config` provided
file. This may also be a curl bug but until that is understood
this commit adds code to double check that if curl returns there
are actually no files left to download. If there are still files
it will error.