sources(curl): use `next` in generated curl config
The curl man page says that almost all options are reset after
a `--next` and that is the desired behavior. It also seems like
two `--cacert` files are possible with `--next` for two different
URLs. This commit uses --next in the config (and also switches
to json --write-out, thanks to Achilleas for the suggestion!).