test: convert `TestStages.run_stage_diff_test()` to pytest
This commit moves the `TestStages.run_stage_diff_test()` and the
`make_stage_tests()` helper into a paramerized pytest.

This makes it more uniform with our other tests that are mostly
pytest based and it also makes it possible to easiyl do a
`pytest --collect-only` to see what is actually run.

Another nice reason to move more to pytest is that it allows
to use custom markers like `@pytest.mark.slow` which we
could use to have a `make quick-test` or something.