spec: remove unneeded dnf depedendency
This was added years ago in 3c19420b3e7786ba87dad15df2cad9de73ef56d3.
Back then, we were constructing the initial buildroot using dnf directly
(stages had network access at that point). We reworked this with the
introduction of the rpm stage, sources and an external depsolver. The
dnf stage itself was actually removed in
892342b978d415a09ae925bb631947f2781be12b. Thus, I think we don't need
to pull dnf anymore, so this commit removes the dependency.