osbuild/solver: provide pacman depsolver
We (used to) have a dependency solver for Pacman and we still have
stages to install packages with Pacman. I'm spending some quality time
trying to build alternative distributions to see what kind of problems I
encounter.

This commit moves the Pacman depsolver into our new solver structure
(e.g. living inside the `osbuild` package, with a separate executable in
`tools/`).

Note that this is broken as we don't have a working Pacman on Fedora 40
or 41 at the moment.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>