osbuild: add new `"build":"org.osbuild.containers-storage:..."`
This commit adds support to construct a build root for a pipeline
from a container instead of a previous pipeline/tree. This is
based on the idea from Ondrej in issue 1804.

A bootc-image-builder manifest would look something like this:
```json
{
  "version": "2",
  "pipelines": [
    {
      "name": "image",
      "build": "org.osbuild.containers-storage:sha256:1234...",
...
   },
   ],
    "sources": {
        "org.osbuild.containers-storage": {
            "items": {
                f"sha256:1234...": {}
            }
        }
    }
}
```

Note that this is just an experiment and needs more thinking how
to abstract/generalize this. It is meant as a faster way to do the
org.osbuild.deploy-container stage does. We should also probably
enforce the uses the container hash instead of a tag when using
"build" for a start.

Closes: https://github.com/osbuild/osbuild/issues/1804