test: extract a common osbuild_fixture() helper
We currently have three copies of the osbuild_fixture() fixture
that are slightly different. This commit extract it into a single
place in `test.py` and shares it accross the various callers.

Note that it would be even nicer to put it into `conftest.py`
but there is a circular import right now when trying this. We
should still do it eventually it just requires a bit more
preperation work.

Split out from https://github.com/osbuild/osbuild/pull/1878