Lint/pylint: update to the latest upstream version v3.3.1
Update pylint to the latest upstream version, to fix a false positive
with Python 3.13, that was resolved in `astroid` dependency in version
`3.3.5`. In order to update the dependency, the `pylint` version itself
needs to be bumped to at least version `3.0.0`.

The error was:
test/run/test_stages.py:21:0: E0611: No name 'Mapping' in module 'collections.abc' (no-name-in-module)

More details: https://github.com/pylint-dev/pylint/issues/10000

Signed-off-by: Tomáš Hozza <thozza@redhat.com>