stages: run useradd, usermod, and passwd with --root
The useradd, usermod, and passwd commands support a `--root` option that
handles chroot-ing for the command.  In general, we prefer using this
option for commands that provide it and relying on the utility itself to
know how to set up the chroot in the way it needs.

The option has been available for these commands since 2011 [1] and it's
unclear why they weren't used originally.

The `mkhomedir_helper` command is still run using our Chroot context, so
the fix introduced in 9071cd0abbf8db059aa34e6cc21612da244e0684 is
unaffected.

[1] https://github.com/shadow-maint/shadow/blob/365279ea95a6c76021e2100e51e71b1991fea32b/ChangeLog#L1339