util/toml: add a TOML writer
None of the Python TOML writers currently used is available in CentOS Stream 10.
The standard Python library supports only TOML readers.  tomli-w
library (https://github.com/hukkin/tomli-w) is endorsed by the Python
documentation as one of the 3rd party TOML writers and it is already used if
present in the system.  Let's copy _writer.py file from tomli-w and use it in
case tomli-w is not present.