Reduce CI to RHEL-9.6 and RHEL-10.0
Reduce CI to run tests only on RHEL 9.6 and 10.0, which both contain
the v141 osbuild, which will need to be patched.

These releases are not yet GA, so keep using the nightly runners.

Do not run the Manifest tests at all in GitLab, because the manifests
from `osbuild/images` `main` branch contain new stage options, which are
not supported by the v141 and thus fail validation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>