Set the release version to 141.2
I forgot to bump the version to 141.1 before making the v141.1 release.
Also, the release action's script that bumps the version did not
support the dot releases. This should be fixed soon.

The version is always set to the "next" to be released version and it is
bumped after we make a release by the GH action.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>