stages/chrony: support specifying refclocks
The refclock directive can be used to specify one or more hardware
reference clocks to be used as a time source.  Each refclock line must
specify a driver and a mandatory parameter, in the form:

  refclock driver parameter

Drivers can have driver-specific options:

  refclock driver parameter:[driver-option,...]

General refclock options can also be specified:

  refclock driver parameter:[driver-option,...] [general-option]...

The stage options schema is written so that the "driver" property is an
object that must match one of four schemas corresponding to the four
drivers, each with a "name" property matching the driver name.
Each driver defines its required property and any optional
driver-specific options.

There are more general refclock options supported than the ones defined
in this commit, but we can add them if and when we need them in the
future.

Note that the restriction on the top-level stage options schema is now
lifted and any set of options can be specified.  Servers are not
required.  However, at least one top-level property is required still.

Docs: https://chrony-project.org/doc/3.4/chrony.conf.html