test: drop `module_platform_id` from depsolve tests
Drop `module_platform_id` as it is now optional and none of
our tests is using it (i.e. has any observable difference if
missing).

Once we start using it we need to add it (and maybe a
"with_platform_id" as parameter on top so that both with/without
platform_id is tested).